requirements.txt
jinja2
tf-keras<=2.15.1
ultralytics
opencv-python
numpy
torch
deepface
pillow
fastapi
